Optimize Datatypes, Errors, Sorting,CollatShares
- Datatypes where not required got reduced in size
- new errors for the Timing Situations, -> They return the blocks to
  show a better error message if a user is not able to reveal yet, or if
  his time already passed. Also for starting a next phase, the error
  message provides information about when the next phase might be
  started.
  - Sorting is also implemented to permute the order when a user is
    allowed to reveal his guess. -> Quicksort
 - Also set a new blockNumber, whenever someone joins, therefore the
   users can't commit multiple guesses with multiple wallets and
   immediately start the game.
 - Collateral Shares are also new, they can be claimed by every
   potential winner. This is the share of all collaterals, that are in
   the contract from users that did not reveal, or revealed the wrong
   guess/salt

+++ backend/contracts/AverageGameFactory.sol
@@ -17,7 +17,7 @@ contract AverageGameFactory {
     function createAverageGame(
         address _address,
         string memory _name,
-        uint256 _maxPlayers,
+        uint16 _maxPlayers,
         uint256 _betAmount,
         uint256 _gameFee,
         uint256 _icon
